Refactor Supabase client to shared module
Moved Supabase client initialization from baseRepository.ts to a new shared module at database/client.ts for better reusability and maintainability. Updated baseRepository.ts to import the client from the new location. Added .gitkeep files to ensure middleware and policies directories are tracked in cxc and web packages.
