+++ packages/server/core/src/repository/baseRepository.ts
@@ -1,27 +1,4 @@
-import { createClient, SupabaseClient } from "@supabase/supabase-js";
-
-const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
-const supabasePublishableKey =
-  process.env.NEXT_PUBLIC_SUPABASE_PUBLISHABLE_DEFAULT_KEY;
-
-if (!supabaseUrl) {
-  throw new Error("Missing env.WEB_SUPABASE_URL");
-}
-
-if (!supabasePublishableKey) {
-  throw new Error("Missing env.WEB_SUPABASE_PUBLISHABLE_KEY");
-}
-
-export const supabase = createClient(supabaseUrl, supabasePublishableKey, {
-  auth: {
-    persistSession: true,
-    autoRefreshToken: true,
-    detectSessionInUrl: true,
-  },
-});
-
-export { createClient } from "@supabase/supabase-js";
-export type { SupabaseClient } from "@supabase/supabase-js";
+import { supabase, SupabaseClient } from "../database/client";
 
 export class BaseRepository {
   protected client: SupabaseClient;
